requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes --format requirements-txt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.2
    # via litellm
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.12.0
    # via
    #   httpx
    #   openai
    #   starlette
attrs==25.4.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via flask
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0 ; platform_python_implementation != 'PyPy'
    # via pynacl
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   flask
    #   functions-framework
    #   litellm
    #   typer-slim
    #   uvicorn
cloudevents==1.12.0
    # via functions-framework
colorama==0.4.6 ; sys_platform == 'win32'
    # via
    #   click
    #   tqdm
deprecation==2.1.0
    # via cloudevents
distro==1.9.0
    # via openai
fastuuid==0.14.0
    # via litellm
filelock==3.20.0
    # via huggingface-hub
flask==3.1.2
    # via functions-framework
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.12.0
    # via huggingface-hub
functions-framework==3.10.0
    # via anisecord
grpcio==1.67.1
    # via litellm
gunicorn==23.0.0
    # via
    #   functions-framework
    #   uvicorn-worker
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
hf-xet==1.2.0 ; platform_machine == 'AMD64' or platform_machine == 'aarch64' or platform_machine == 'amd64' or platform_machine == 'arm64' or platform_machine == 'x86_64'
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   huggingface-hub
    #   litellm
    #   openai
huggingface-hub==1.2.1
    # via tokenizers
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==8.7.0
    # via litellm
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via
    #   flask
    #   litellm
jiter==0.12.0
    # via openai
jsonschema==4.25.1
    # via litellm
jsonschema-specifications==2025.9.1
    # via jsonschema
litellm==1.80.9
    # via anisecord
markupsafe==3.0.3
    # via
    #   flask
    #   jinja2
    #   werkzeug
multidict==6.7.0
    # via
    #   aiohttp
    #   yarl
openai==2.9.0
    # via litellm
packaging==25.0
    # via
    #   deprecation
    #   gunicorn
    #   huggingface-hub
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
pycparser==2.23 ; implementation_name != 'PyPy' and platform_python_implementation != 'PyPy'
    # via cffi
pydantic==2.12.5
    # via
    #   litellm
    #   openai
pydantic-core==2.41.5
    # via pydantic
pynacl==1.6.1
    # via anisecord
python-dotenv==1.2.1
    # via litellm
pyyaml==6.0.3
    # via huggingface-hub
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2025.11.3
    # via tiktoken
requests==2.32.5
    # via
    #   anisecord
    #   tiktoken
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
shellingham==1.5.4
    # via huggingface-hub
sniffio==1.3.1
    # via openai
starlette==0.50.0
    # via functions-framework
tiktoken==0.12.0
    # via litellm
tokenizers==0.22.1
    # via litellm
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   openai
typer-slim==0.20.0
    # via huggingface-hub
typing-extensions==4.15.0
    # via
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
    #   typer-slim
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
urllib3==2.6.1
    # via requests
uvicorn==0.38.0
    # via
    #   functions-framework
    #   uvicorn-worker
uvicorn-worker==0.4.0
    # via functions-framework
watchdog==6.0.0
    # via functions-framework
werkzeug==3.1.4
    # via
    #   flask
    #   functions-framework
yarl==1.22.0
    # via aiohttp
zipp==3.23.0
    # via importlib-metadata
